Rename data-parse to create-contact.
