+++ src/content/plugins/linkedin.js
@@ -7,7 +7,7 @@ import {isQuill} from '../editors/editor-quill.js';
 import {insertTemplate} from '../editors/editor-universal.js';
 import {insertContentEditableTemplate} from '../editors/editor-contenteditable.js';
 import {htmlToText} from '../utils/plain-text.js';
-import {createContact} from '../utils/data-parse.js';
+import createContact from '../utils/create-contact.js';
 import {enableBubble} from '../bubble.js';
 
 function before (params, data) {
